moves data and components to their own files
