+++ data/jedi.js
@@ -1,0 +1,56 @@
+import tempJediCard from '../public/images/jedi/allyship-space-temp.png';
+
+const allyshipSpaces = [
+  {
+    title: 'jedi allyship space',
+    date: 'April 20th 2021', // TODO: timestamp + conversions
+    location: 'Boelter 2763',
+    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    pic: tempJediCard,
+    alt: 'a photo of Helia smiling',
+    rsvp: 'https://medium.com/acm-at-ucla/may-the-force-be-with-acm-at-ucla-49e9258d71e1',
+    slides: 'https://medium.com/acm-at-ucla/may-the-force-be-with-acm-at-ucla-49e9258d71e1',
+  },
+  {
+    title: 'another jedi allyship space',
+    date: 'April 20th 2021', // TODO: timestamp + conversions
+    location: 'Boelter 2763',
+    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
+    pic: tempJediCard,
+    alt: 'a photo of Helia smiling',
+    slides: 'https://medium.com/acm-at-ucla/may-the-force-be-with-acm-at-ucla-49e9258d71e1',
+  },
+];
+
+const jedis = [
+  {
+    name: 'helia woo',
+    pronouns: 'she/her',
+    bio: 'I lead JEDI!',
+    pic: 'https://teachla.uclaacm.com/img/team/hwoo.png',
+    alt: 'a photo of Helia smiling',
+  },
+  {
+    name: 'helia woo',
+    pronouns: 'she/her',
+    bio: 'I lead JEDI!',
+    pic: 'https://teachla.uclaacm.com/img/team/hwoo.png',
+    alt: 'a photo of Helia smiling',
+  },
+  {
+    name: 'helia woo',
+    pronouns: 'she/her',
+    bio: 'I lead JEDI!',
+    pic: 'https://teachla.uclaacm.com/img/team/hwoo.png',
+    alt: 'a photo of Helia smiling',
+  },
+  {
+    name: 'helia woo',
+    pronouns: 'she/her',
+    bio: 'I lead JEDI!',
+    pic: 'https://teachla.uclaacm.com/img/team/hwoo.png',
+    alt: 'a photo of Helia smiling',
+  },
+];
+
+export { allyshipSpaces, jedis };
